1. Created a "types" folder in the "src" directory and relocated the "Enums.ts" file to it. 2. Renamed the "enums" file to "Enums". 3. Refactored folder names to use lowercase. 4. Introduced a "helper.ts" file containing the "getColumns" function to separate logic from JSX. 5. Refactored data fetching from a file following a provided example.
